+++ geotrek/tourism/parsers.py
@@ -15,6 +15,7 @@ from geotrek.common.parsers import (
     AttachmentParserMixin,
     GeotrekParser,
     LEIParser,
+    OpenStreetMapAttachmentsParserMixin,
     OpenStreetMapParser,
     Parser,
     TourInSoftParser,
@@ -1388,7 +1389,9 @@ class GeotrekInformationDeskParser(GeotrekParser):
         )
 
 
-class InformationDeskOpenStreetMapParser(OpenStreetMapParser):
+class InformationDeskOpenStreetMapParser(
+    OpenStreetMapAttachmentsParserMixin, OpenStreetMapParser
+):
     """Parser to import information desks from OpenStreetMap"""
 
     type = None
@@ -1410,7 +1413,6 @@ class InformationDeskOpenStreetMapParser(OpenStreetMapParser):
     natural_keys = {
         "type": "label",
     }
-    non_fields = {}
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -1430,8 +1432,10 @@ class InformationDeskOpenStreetMapParser(OpenStreetMapParser):
         if type == "node":
             geom = Point(float(lng), float(lat), srid=self.osm_srid)  # WGS84
             geom.transform(settings.SRID)
-            return geom
         elif type == "way":
-            return self.get_centroid_from_way(area)
+            geom = self.get_centroid_from_way(area)
         elif type == "relation":
-            return self.get_centroid_from_relation(bbox)
+            geom = self.get_centroid_from_relation(bbox)
+
+        geom.transform(settings.SRID)
+        return geom